Updated pkg structure: moved UserNotFoundException to Advice
